more clean up; move parse_entry_point into _utils.py
