+++ babelizer/_files/init_py.py
@@ -5,16 +5,14 @@ from collections.abc import Mapping
 from typing import Any
 
 
-def render(plugin_metadata: Mapping[str, Any]) -> str:
+def render(context: Mapping[str, Any]) -> str:
     """Render __init__.py."""
-    package_name = plugin_metadata["package"]["name"]
+    package_name = context["package"]["name"]
 
     imports = [f"from {package_name}._version import __version__"]
-    imports += [
-        f"from {package_name}._bmi import {cls}" for cls in plugin_metadata["library"]
-    ]
+    imports += [f"from {package_name}._bmi import {cls}" for cls in context["library"]]
 
-    names = [f"    {cls!r},".replace("'", '"') for cls in plugin_metadata["library"]]
+    names = [f"    {cls!r},".replace("'", '"') for cls in context["library"]]
 
     return f"""\
 {os.linesep.join(sorted(imports))}
